requirements.txt
bs4
halo
openai
pypdf2
pyyaml
